Organizando melhor as funções de utilidade
