+++ test/main.py
@@ -1,6 +1,11 @@
 """Arquivo que deve ser executado para rodar os testes"""
 import unittest
-from utils import ChiSquareTest, CrammerVTest, ContigencyCoefficientTest
+import os
+import sys
+sys.path.append(os.getcwd())
+
+# Testes das funções estatísticas
+from test.utils.statistic import ChiSquareTest, CrammerVTest, ContigencyCoefficientTest
 
 
 if __name__ == "__main__":
